+++ testdata/regression/provided_helper/provided.go
@@ -4,5 +4,5 @@ package wasm2go
 // code never emits (it has no i64.load); the translator must emit the helper
 // because this file references it.
 func (m *Module) _peek(addr int32) int64 {
-	return int64(load64(m.memory, uint32(addr)))
+	return int64(load64(m.memory, uint64(uint32(addr))))
 }
